Move Concern to root namespace
