--- sig/kangaru/concerns/concern.rbs
@@ -1,15 +1,0 @@
-module Kangaru
-  module Concerns
-    module Concern : Module
-      @included: untyped
-
-      def append_features: (untyped) -> void
-
-      def class_methods: { [self: singleton(Class)] -> void } -> void
-
-      def included: (?untyped?) ?{ -> void } -> void
-
-      def evaluate_concern_blocks!: (untyped) -> void
-    end
-  end
-end
